fix: training

+++ nfm/data/efd.py
@@ -3,10 +3,10 @@ from numpy.typing import NDArray
 
 
 def _phase_shift_efd(
-    coeffs: NDArray[np.float32],
-    theta: NDArray[np.float32],
-    harmonic_indices: NDArray[np.float32],
-) -> NDArray[np.float32]:
+    coeffs: NDArray[np.float64],
+    theta: NDArray[np.float64],
+    harmonic_indices: NDArray[np.float64],
+) -> NDArray[np.float64]:
     """Apply a start-point phase shift θ to each harmonic."""
     cos_terms = np.cos(harmonic_indices * theta[:, None])
     sin_terms = np.sin(harmonic_indices * theta[:, None])
@@ -22,15 +22,15 @@ def _phase_shift_efd(
 
 
 def normalize_efd_for_starting_point(
-    coeffs: NDArray[np.float32],
-) -> NDArray[np.float32]:
+    coeffs: NDArray[np.float64],
+) -> NDArray[np.float64]:
     """Rotate EFD coefficients to eliminate contour start-point dependence."""
-    mask = np.linalg.norm(coeffs[:, 0], axis=1) > np.finfo(np.float32).eps
+    mask = np.linalg.norm(coeffs[:, 0], axis=1) > np.finfo(np.float64).eps
     if not np.any(mask):
         return coeffs
 
     a1, b1, c1, d1 = coeffs[mask, 0].T
-    harmonic_indices = np.arange(1, coeffs.shape[1] + 1, dtype=np.float32)
+    harmonic_indices = np.arange(1, coeffs.shape[1] + 1)
     theta = 0.5 * np.arctan2(2.0 * (a1 * b1 + c1 * d1), a1**2 - b1**2 + c1**2 - d1**2)
 
     rotated = _phase_shift_efd(coeffs[mask], theta, harmonic_indices)
@@ -49,8 +49,8 @@ def normalize_efd_for_starting_point(
 
 
 def elliptic_fourier_descriptors(
-    contour: NDArray[np.float32], order: int
-) -> NDArray[np.float32]:
+    contour: NDArray[np.float64], order: int
+) -> NDArray[np.float64]:
     """Computes the Elliptic Fourier Descriptors for a set of contours.
 
     Args:
